pw_log_basic: Add printf format checking
Use PW_PRINTF_FORMAT and fix an issue it exposes.

+++ pw_kvs/public/pw_kvs/flash_memory.h
@@ -319,7 +319,7 @@ class FlashPartition {
   }
 
  protected:
-  Status CheckBounds(Address address, uint32_t len) const {
+  Status CheckBounds(Address address, size_t len) const {
     if (address + len > GetSizeBytes()) {
       PW_LOG_ERROR(
           "Attempted out-of-bound flash memory access (address: %" PRIu32
